manager -> controller and other trivial renaming.
